fix(maps): draw place photos and centre the pin on the Google map
Two defects in the Google renderer, both visible next to the same trip on
the Leaflet map.

Pins were drawn without a photo. Reading place.image_url is not enough to
fix it: most places carry no image of their own, and the picture on a pin
comes from the photo service's runtime cache keyed by provider id or
coordinates. That lookup lived inside MapViewGL, so it is lifted into a
usePlacePhotos hook that both renderers share rather than copied — the
marker effect now redraws as thumbnails arrive.

Pins also sat half their own height above the place they marked. Google
anchors an AdvancedMarkerElement by the bottom edge of its content, while
createMarkerElement draws a pin meant to be anchored at its centre, which
is what the GL renderers ask for with anchor: 'center'. The offset also
moved a pin whenever its size changed, since selection grows it from 36px
to 44px.
